refactor: Adjusts search_deployment schema definitions with structure of new schema scaffolding command
