+++ internal/service/searchdeployment/tfplugingen/generator_config.yml
@@ -1,0 +1,22 @@
+provider:
+  name: mongodbatlas
+
+resources:
+  search_deployment:
+    read:
+      path: /api/atlas/v2/groups/{groupId}/clusters/{clusterName}/search/deployment
+      method: GET
+    create:
+      path: /api/atlas/v2/groups/{groupId}/clusters/{clusterName}/search/deployment
+      method: POST
+    schema:
+      ignores: []
+
+
+data_sources:
+  search_deployment:
+    read:
+      path: /api/atlas/v2/groups/{groupId}/clusters/{clusterName}/search/deployment
+      method: GET
+    schema:
+      ignores: []
